+++ src/fpsgame/ctf.h
@@ -363,7 +363,7 @@ struct ctfclientmode : clientmode
         drawblip(d, x, y, s, flagblip ? (f.owner ? f.owner->o : (f.droptime ? f.droploc : f.spawnloc)) : f.spawnloc, flagblip);
     }
 
-    int clipconsole(int w, int h)
+    float clipconsole(float w, float h)
     {
         return (h*(1 + 1 + 10))/(4*10);
     }
